[deckhand] Remove gabbi functional result reporting stage
This removes the stage from the Deckhand Jenkins file responsible for
publishing the results of the gabbi functional test run generated
by the functional test stage. This is because of a conflict of
competing interests in Deckhand:

  * Deckhand uses pytest-html plugin for generating the HTML report
    for functional tests
  * But gabbi fixtures don't work with pytest [0]
  * But Deckhand should use gabbi fixtures to have better testing
  * There aren't any decent non-pytest html plugins
  * Therefore Deckhand should just not bother with the html report

Besides that, the HTML report is publicized and no one really looks
at it so it can be safely removed.

[0] cf. note under http://gabbi.readthedocs.io/en/latest/fixtures.html?highlight=pytest#inner-fixtures

Change-Id: I7cd4bcab8682f502f8fded3e08eefa142ae57a10