Iterate on shared libs, global vars
This PS aims to iterate on and improve the shared libs, as well
as add in some global vars. The biggest thing is to generalize
what some of these component related functions do and to not
make assumptions about what the consumer of said function wants
to do with it. Instead, it looks to return back to them everything
it gets from the component and then lets them make decisions on
what they want to use, resiliency-related features, etc. In addition,
simplification is key.

Change-Id: I7094b163fc240799de1c39caa00773dac0d85d6a