Add Dockerfile for base images
The Dockerfile allow to control exectly what software
is added to the images of the containers.

The following are rebuild from scratch in the order:
- docker-brew-ubuntu-core/xenial
- buildpack-deps/xenial/curl
- buildpack-deps/xenial/scm
- buildpack-deps/xenial
- python/3.5/xenial
- python/3.6/xenial

The next step will be to remove all the unused packages
and to let everythin point internal in artifactory.

Committed on behalf of Jerome Brette

Change-Id: If873765187d00cf42e0a9cb8bff2ff9fc009badc