Update patch set 4
Patch Set 4:

(2 comments)

> We should extend this job's functionality to run at given intervals
 > and automatically cleanup slaves/stacks that are over a certain
 > age.
That's great suggestion and that wan in my proposal for the user story, but it hasn't been included into the user story. And to implement it I would need to play a bit with heat commands to understand how to get age of a VM.

Thanks for good comment on EXCLUDE_NODE_LIST.contains(xxx).

Patch-set: 4