Shipyard shared lib - allow intermediate commits
This PS adds the "allow-intermediate-commits" flag into the Shipyard
action creation as a means of sidestepping a more recent issue with
how Shipyard is determining the current document revision vs. what's
actually in Deckhand. I presume this will be able to be reversed
in the near future

Change-Id: I690c2eb26971a479f8683bdbf04e9196464813df
See: TDP BU438478