[456627] add doc publishing support
This change aims to add a doc publish to artifactory for aic-aqa-*
projects so that we can move this job to NC cicd and support UML
diagrams in our docs.

upon a succesful code-review, the job will attempt to publish to
the "aqua-docs" repository with the name of the project without the
"aic-aqa-" prefix as the containing folder for the project. So for
example, aic-aqa-api would be stored in the folder aqua-docs/api/
and aic-aqa-infrastructure-sonobuoy/ would be stored in
aqua-docs/infrastrcuture-sonobuoy/

Change-Id: I450d2c13ca8b967e4baab0eb5df5c1cab7ac3d7a