add deleteArtifacts method
This commit aims to add a method to delete artifacts from artifactory.

In order to properly publish documentation upon patchsets being
created or merged, a deleteArtifacts method is needed to upload files
because there is no way to delete files with the putArtifacts method.

After calling the deleteArtifacts method, the output will show the
response code from the curl command. A response code of 404 indicates
that the user is trying to delete a file that does not exist or the user
does not have permission to delete such a file.

Change-Id: I45a6be54ac972f26a8b0bcfe9d9201ee1b997912