Mount ssh-agent socket to loci container during docker run
That is the fourth change from chain, that intends to create an ability
for loci to use sources that require authentication while creating
docker images. With this change internal dependencies may be specified
in upper-constraints.txt.

With this change ssh-agent socket, available on host, is reused in
loci container during docker run. Also SSH_DATA internal variable is
introduced as a source data to compile loci docker ssh config and what
credentials to use in ssh-agent.

Change-Id: I9a6524cedd0ab7a0a7117f227c8055ca4d95f1ba