add deleteArtifacts method
This commit aims to add a method to delete artifacts from artifactory.

In order to properly publish documentation upon patchsets being
created or merged, a deleteArtifacts method is needed to upload files
because there is no way to delete files with the putArtifacts method.

Attempting to delete 1 artifacts.
[Pipeline] httpRequest
[Pipeline] echo
The API status code 204 was received from artifactory after the delete request.
[Pipeline] httpRequest
[Pipeline] echo
A total of 0 artifacts currently match the given path test-repo/hello.txt after the delete request.

Change-Id: I45a6be54ac972f26a8b0bcfe9d9201ee1b997912