WIP: add deleteArtifacts method
This commit aims to add a method to delete artifacts from artifactory.

In order to properly publish documentation upon patchsets being
created or merged, a deleteArtifacts method is needed to upload files
because there is no way to delete files with the putArtifacts method.

I'm working on using groovy instead of shell.

Example output is coming soon..

Change-Id: I45a6be54ac972f26a8b0bcfe9d9201ee1b997912