Remove credential names and mention of aqua
This patch attempts to remove exact names of credentials
or folders in artifactroy from the vars directory from comments
and where ever else is possible.

Methods that use the creds as a required parameter were not
modified, however, methods that set creds as a default were changed
for creds to be a required parameter.

Change-Id: I22a8411f80d65fc32ac1b593a53256492c5f1424