Increaes retry count in API calls to Keystone/Shipyard
This PS increases the number of tries to attempt to retry when
making Keystone and/or Shipyard API calls and a failure is
encountered. The thought is if one of these services is mid-upgrade,
we want to provide amply opportunity for it to recover and not
fail the pipeline prematurely.

Change-Id: Id4c576e73aa510ad283930eaf60a5cb0fd2461cc