mini-mirror: Restrict image job to master branch
Currently, the mini-mirror build job runs on every branch of
aic-clcp-manifests; however, the nc2/undercloud branch uses a new job
that is no longer compatible with the job in this repository.

This change restricts the job to the master branch, so it will not vote
negatively on the changes to other branches.

Once the nc2/undercloud branch is merged to master, this pipeline will
be permanently disabled.

Signed-off-by: Drew Walters <andrew.walters@att.com>
Change-Id: Ia4e16dd5e6be0c6b5f60eed11a9347a38654e7df