Fix submitPatchset method to follow doc string
Currently gerrit.submitPatchset has "refspec" defined in its doc string
without actually defining or using such a parameter, and it also has
the "repoName" parameter without defining it in the doc string. This
updates the method to resolve this.

The addition of the refspec parameter is especially useful for changes
that are not being pushed to the master branch of a repository,
although "refs/for/master" has been left as the default value to avoid
problems with existing usage of the method.

Change-Id: Ia5f0b1a26cbd8b8e4f0c41e0c0f055fb0d9c5721