Move the execution of find seed
Moves the execution of the find_seed command to follow the check for
patchset type, preventing a spurious error when more than one
seed*.groovy file exists in the same directory as the change.

Because the script short circuits if the change is a patchset-created
(and does nothing with the changed file, this only removes the spurious
error, and does not impact the rest of the logic.

Change-Id: I325062af6d6b42b6cbb3188ab4599a3127bb0c4c